feat: increment parser
