+++ src/bsl/codeModel/model/bslCodeModel.ts
@@ -1,6 +1,6 @@
 import { BaseSymbol, CompositeSymbol, Method } from "@/common/codeModel";
 import { VariablesScope } from "./interfaces";
-import { MethodsCalculator, ModelCalculator, TypesCalculator, VariablesCalculator } from "../calculators";
+import { MethodsCalculator, TypesCalculator, VariablesCalculator } from "../calculators";
 import { FunctionDefinitionSymbol, ProcedureDefinitionSymbol } from "./definitions";
 import { ParentsCalculator } from "../calculators";
 import { Emitter, IEvent } from "monaco-editor-core";
@@ -8,7 +8,11 @@ import { AutoDisposable } from "@/common/utils/autodisposable";
 import { BslVariable } from "./variables";
 
 export class BslCodeModel extends AutoDisposable implements VariablesScope, CompositeSymbol {
-    calculators: ModelCalculator[] = [new ParentsCalculator(), new VariablesCalculator(), TypesCalculator.instance]
+    calculators = {
+        parents: new ParentsCalculator(),
+        variables: new VariablesCalculator(),
+        types: TypesCalculator.instance
+    }
     children: BaseSymbol[] = []
     vars: BslVariable[] = []
 
@@ -26,8 +30,16 @@ export class BslCodeModel extends AutoDisposable implements VariablesScope, Comp
         return this.children
     }
 
-    afterUpdate() {
-        this.calculators.forEach(c => c.calculate(this))
+    async afterUpdate(symbol: BaseSymbol | BslCodeModel) {
+        if (symbol instanceof BslCodeModel) {
+            this.calculators.parents.calculate(this)
+            this.calculators.variables.calculate(this)
+            await this.calculators.types.calculate(this)
+        }else{
+            this.calculators.parents.calculate(this)
+            // this.calculators.variables.calculate(this)
+            // await this.calculators.types.calculate(this)
+        }
         this.onDidChangeModelEmitter.fire(this)
     }
